web: Make sidebar list shorter
Due to redundant whitespace, HTML of versions list with all Linux
tags (including tags that are normally filtered away) takes around
2 megabytes.

This commit aims to reduce that number by removing all whitespace from
the sidebar. To reduce the size of the sidebar further, tags of the
lowest level of the version tree were simplified.
This cut the size of the versions list HTML to around 300 kilobytes.

Large sidebar HTML is a problem - the version list on the official
instance is so large, that the subtrees for versions v2-v0 aren't
visible immediately after opening the page. Since the sidebar is
now above content, the content does not load until the sidebar loads.

Further improvement could be made by simplifying links that redirect to
a different repostitory version. For example, a short query parameter
could be introduced. If this parameter would be present in the URL,
backend could generate a redirect to URL with a changed version. Version
links could then just contain a relative URL with the paramenter,
instead of the absolute URL that grows the with the file path.