opensearch: Add opensearch manifest
OpenSearch is a standard that allows browsers to discover search engines.
A special XML file with information about the search engine is linked
from its webpage in a link tag.
OpenSearch is supported by Firefox and (maybe) macOS Safari.
It used to be supported by Chromium-based browsers, but that feature was
deprecated.
https://stackoverflow.com/questions/56400952/does-chrome-allow-auto-discovery-of-opensearch

MDN Docs:
https://developer.mozilla.org/en-US/docs/Web/XML/Guides/OpenSearch

OpenSearch specification:
https://github.com/dewitt/opensearch/blob/master/opensearch-1-1-draft-6.md